Implement TextBuffer in terms of a stack of composed patches

+++ src/core/patch.cc
@@ -1196,7 +1196,7 @@ std::string Patch::get_json() const {
 size_t Patch::get_change_count() const { return change_count; }
 
 void Patch::clear() {
-  delete_node(&root);
+  if (root) delete_node(&root);
 }
 
 void Patch::rebalance() {
